feat(1.21.1): 移植至 Minecraft 1.21.1（malilib 0.21.10 + litematica 0.19.61）
基于 support/1.21.11 分支下调至 1.21.1 旧渲染管线世代。所有 API 签名
经 javap 字节码核验，非版本号猜测。

构建配置：
- gradle.properties: minecraft 1.21.1 / loader 0.16.10 / fabric-api 0.116.14+1.21.1
- build.gradle: Loom 1.15.4，依赖全走 Modrinth maven（malilib GcWjdA9I:C99LEy6r /
  litematica bEpr0Arc:o7i8oi1Z），删除 libs/ 本地 jar
- fabric.mod.json: minecraft >=1.21.1 <1.21.4

渲染管线（LitematicaBridge）：
- IRenderer.onRenderWorldLast(Matrix4f, Matrix4f) 仅 2 参数（1.21.11 为 7 参数）
- RenderUtils.renderAreaOutline/renderBlockOutline 需额外传 Minecraft 参数
- Color4f 用 util.Color4f（旧路径），非 util.data.Color4f
- writeToFile(File,...) 而非 Path，调用处 dir.toFile()
- BoxType 在 0.19.61 为 package-private，无法用 renderSelectionBox，改回 RenderUtils

GUI 鼠标事件（screen/widget 全量适配）：
- 移除 MouseButtonEvent/KeyEvent（1.21.5+ 专有），改 int 原始参数
- onMouseClicked(int,int,int) / onMouseScrolled(int,int,double,double)
- GuiContext 不存在，drawContents/drawScreenBackground 改用 GuiGraphics
- keyPressed(KeyEvent) → onKeyTyped(int,int,int)
- onMouseDragged 不存在于 GuiBase，WeCommandScreen 改用 Screen.mouseDragged
- WeGuiConfigs 移除 IConfigGuiAllTab（0.21.10 无此接口）
- wrapper.textField() → getTextField()；screen.font → screen.textRenderer

Mixin：
- ButtonBaseMixin.onMouseScrolledImpl 改 int+int+double+double
- ClientPacketListenerMixin 移除 m_246046_/m_246039_ SRG 备选名

其余：
- AxeModeHandler: Identifier→ResourceLocation，UseItemCallback 返回 InteractionResultHolder
- ModeIndicator: PoseStack pushPose/translate3参/scale3参/popPose
- WeCommands: biome 走动态注册表 BuiltInRegistries.REGISTRY.get(Registries.BIOME.location())

编译通过 + jar 产出 + mixin remap 无警告。运行时验证待用户本地手动进行。

+++ src/client/java/com/sow/wegui/client/mixin/ButtonBaseMixin.java
@@ -15,7 +15,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public class ButtonBaseMixin {
 
     @Inject(method = "onMouseScrolledImpl", at = @At("HEAD"), cancellable = true)
-    private void wegui$disableScrollClick(double mouseX, double mouseY, double scrollX, double scrollY,
+    private void wegui$disableScrollClick(int mouseX, int mouseY, double scrollX, double scrollY,
                                           CallbackInfoReturnable<Boolean> cir) {
         cir.setReturnValue(false);
     }
